% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cliDat}
\alias{cliDat}
\title{Survival Data}
\format{
A data frame with 368 rows and 2 variables:
\describe{
  \item{time}{survival time(day)}
  \item{status}{survival status(0=Dead,1=Alive)}
}
}
\usage{
cliDat
}
\description{
This data contains 368 samples survival data, including time and status.
}
\keyword{datasets}
