% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprset}
\alias{exprset}
\title{Expression matrix of 7 genes in 368 samples}
\format{
A data frame with 368 rows and 7 variables
}
\usage{
exprset
}
\description{
This data contains 7 markers, "TREM1", "EXO1", "COLEC10", "C7", "FCN3", "DEFA6", "KLK7",
   expression in 368 samples.
}
\keyword{datasets}
