% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.data.R
\name{glmnet.data}
\alias{glmnet.data}
\title{To Acquire The Glmnet Data for Visualization}
\usage{
glmnet.data(fit, xvar = c("norm", "lambda", "dev"))
}
\arguments{
\item{fit}{The result of \link{glmnet}.}

\item{xvar}{Observation type, which will be x-axis mapping.
You can choose "norm","lambda" or "dev".}
}
\value{
A list about the data of glmnet
}
\description{
Return a list, which contains the basic data needed for drawing, 
    to customize the drawing.
}
