% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmplot.R
\name{glmplot}
\alias{glmplot}
\title{Visualization of Lasso}
\usage{
glmplot(fit, xvar, mode, colours = NULL)
}
\arguments{
\item{fit}{The result of \link{glmnet}.}

\item{xvar}{Observation type, which will be x-axis mapping.
You can choose "norm","lambda" or "dev".}

\item{mode}{A number of Figure mode. 1 or 2.}

\item{colours}{Vector of colours to use for n-colour gradient.
see \link{scale_color_gradientn}}
}
\value{
ggplot objest
}
\description{
Using ggplot2 to plot more elegant figure.
}
\examples{
\dontrun{
library(survival)
library(glmnet)
library(glmplot)
data(exprset)
data(cliDat)
x <- as.matrix(exprset)
y <- as.matrix(Surv(cliDat[,1],cliDat[,2]))
fit <- glmnet(x,             
              y,             
              family = "cox",  
              alpha = 1,      
              nlambda = 1000)   

glmplot(fit,xvar = "dev",mode = 1,
   colours=c("#3B4992FF","#EE0000FF","#008B45FF","#631879FF","#008280FF","#BB0021FF"))
}
}
